+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ## [unreleased]
 
  * Portal: Load the JS resources for apps sequentially, because if you have more than one bundle they typically depend on each other
+ * LDAP Security Provider: Add all roles from the *groupToRoleMapping* and *userToRoleMapping*, to improve the autocomplete in the Admin UI
+ * Simple Security Provider: Add all roles from *users.json* to the known roles, to improve the autocomplete in the Admin UI
 
 ## 1.6.4 (February 1, 2021)